agents/TradyyyBot/requirements.txt
flask>=2.3,<3.1
click>=8.1.7,<9
pandas>=2.0
numpy>=1.24
yfinance>=0.2.40
matplotlib>=3.7
requests>=2.31
urllib3>=2.0
python-dateutil>=2.8.2
google-genai>=1.0.0
